ACCUMULO-2815 Stub out an invocation handler to update the principal from SASL into the tcredentials.
I think this should be an appropriate way to ensure that clients can't attempt to act
as a user other than themself. We can override whatever was passed in via the RPC,
rely on SASL/GSSAPI giving us the proper principal from kerberos and using that
as the principal for authorization (and skipping more authentication later).